Blurred component: Added common useBlurredState hook for managing isBlurred state & minHeight  logic on both web & native
